Move channel ordering to the channels view
