--- packages/ui/src/ConfigurationsView/ChannelOrderRow/index.module.scss
@@ -1,61 +1,0 @@
-.items {
-  margin-top: 0.5rem;
-}
-
-@media (min-width: 768px) {
-  .items {
-    column-count: 2;
-  }
-}
-
-.item {
-  align-items: center;
-  display: flex;
-  padding: 0.25rem 0;
-}
-
-.item:last-child {
-  margin-bottom: 0;
-}
-
-.label {
-  align-items: center;
-  cursor: grab;
-  display: flex;
-  font-size: 14px;
-  gap: 0.25rem;
-  font-weight: 500;
-}
-
-.drop {
-  background: var(--color-blue-50);
-}
-
-.channel {
-  align-items: center;
-  display: flex;
-  gap: 0.25rem;
-}
-
-.channel::before {
-  background-color: var(--color-text);
-  content: '';
-  mask-size: contain;
-  width: 14px;
-  height: 14px;
-}
-
-.channel.lock::before {
-  mask-image: url('https://www.linen.dev/icons/lock.svg');
-}
-
-.channel.hash::before {
-  mask-image: url('https://www.linen.dev/icons/hash.svg');
-}
-
-.channelNumber {
-  font-size: 0.75rem /* 12px */;
-  line-height: 1rem /* 16px */;
-  font-style: italic;
-  color: var(--color-gray-400);
-}
